requirements.txt
attrs==23.1.0
blinker==1.6.2
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
configparser==6.0.0
decorator==5.1.1
exceptiongroup==1.1.1
Flask==2.3.2
idna==3.4
iiif-prezi3==1.2.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
ply==3.11
pydantic==1.10.10
pytest==7.4.0
python-dateutil==2.8.2
requests==2.31.0
requests-cache==1.1.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.0
url-normalize==1.4.3
urllib3==2.0.3
Werkzeug==2.3.6
Whoosh==2.7.4
